backend/requirements.txt
fastapi>=0.111,<1
uvicorn>=0.30,<1
chromadb>=0.5,<1
sentence-transformers>=3,<4
PyMuPDF>=1.24,<2
openai-whisper>=20231117
transformers>=4.42,<5
torch>=2.3,<3
Pillow>=10,<11
icalendar>=5,<6
python-docx>=1.1,<2
pytesseract>=0.3,<1
aiosqlite>=0.20,<1
httpx>=0.27,<1
pydantic>=2.7,<3
python-multipart>=0.0.9,<1

# Test dependencies
pytest>=8.2,<9
pytest-asyncio>=0.23,<1
pytest-cov>=5,<6
